ENH: more cleanup
